Rename(add) : Add 내 index 제거
