--- src/components/ProductionAdd/ItemAdd/index.style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const SubjectWrapper = styled.h3`
-  font-weight: 900;
-  font-size: 20px;
-  margin: 20px 10px;
-`;
-
-export const ItemSelect = styled.select`
-  width: 300px;
-  height: 30px;
-`;
