+++ src/pages/productions/add.js
@@ -1,4 +1,5 @@
-import ProductionAdd from 'components/ProductionAdd';
+import ProductionAdd from "components/ProductionAdd/ProductionAddPage";
+
 
 export default function Add() {
   return (
